fix(preempt,vm): CUE-conversion follow-up + qemu-shutdown deletion + TOCTOU restore-race fix
Three follow-up items surfaced during PR review, all folded into this
cutover per the standing operator directives (proven-dead code deletes
where found; any wire type a PR touches converts to CUE now):

1. OwnerPID int64 widening: the paired sdk PR's arbiter.cue conversion
   generates PreemptLease.OwnerPID as int64 (CUE int has no narrower-width
   annotation, was hand-written int before). Fixes every construction site
   (arbiter.go's two os.Getpid() assignments) and the ownerAlive/
   procStartTime signatures in arbiter_support.go, plus the matching test
   fixtures in arbiter_test.go.

2. Deletes the now-fully-senderless "qemu-shutdown" VmOp case in
   candy/plugin-vm/provider.go (confirmed via git grep 'VmPluginEnv{' —
   zero remaining setters tree-wide) rather than registering it for a
   later batch: qemuForceShutdown/qemuGracefulShutdown themselves stay
   (vm.go's own direct-QEMU CLI path calls them live), the now-dead
   StateDir field on provider.go's local vmEnv struct is removed, and 3
   stale plugin.go doc-comment mentions of "qemu-shutdown" are swept.

3. Fixes a TOCTOU restore race the validator's independent bed re-runs
   caught: pluginHolderStart's VM branch was check-then-act
   (pluginHolderExists's domain-state RPC, then startVMPlugin's separate
   start RPC) with no atomicity between them — the bed's own
   cleanup-members/rebuild-members-down teardown can destroy the domain in
   that window, producing a nondeterministic "domain not found" hard error
   (lease stranded) on some runs and a clean departed-holder no-op on
   others, same code, same scenario. Eliminates the window instead of
   synchronizing it (R4): attempts the start unconditionally and folds
   departed-holder detection into the start RPC's own error reply
   (isDomainNotFoundErr), matching candy/plugin-vm's actual
   "domain not found: ..." reply text. The now-redundant VM-branch
   existence pre-check is deleted; pluginHolderExists keeps only the
   pod-venue check. New break-it-proven regression tests using a fake
   pb.ExecutorServiceClient (mirroring candy/plugin-deploy-vm/
   lifecycle_test.go's established pattern): TestIsDomainNotFoundErr,
   TestPluginHolderStart_VmDomainNotFoundIsNoOp,
   TestPluginHolderStart_VmOtherErrorPropagates.

Also repoints all 3 gitlinks to their latest merged commits: sdk to
66bb8ed (the CUE-conversion PR's tip, go.mod bumped to v0.2026203.1028),
box/fedora to 01331c7 (distro-fedora#6, the check-preempt-vm-live bed now
on fedora main), plugins to 3d4d67c (the docs/lessons batch).

+++ candy/plugin-preempt/arbiter.go
@@ -191,7 +191,7 @@ func (a *ResourceArbiter) AcquireExclusive(claimant string, tokens []string, cla
 		Transient:  transient,
 		Preempted:  toStop,
 		Created:    a.nowUTC(),
-		OwnerPID:   os.Getpid(),
+		OwnerPID:   int64(os.Getpid()),
 		OwnerStart: selfProcStart(),
 	}
 	ledger.Leases = append(ledger.Leases, lease)
@@ -269,7 +269,7 @@ func (a *ResourceArbiter) AcquireShared(claimant string, tokens []string, claimA
 		Transient:  transient,
 		Preempted:  toStop,
 		Created:    a.nowUTC(),
-		OwnerPID:   os.Getpid(),
+		OwnerPID:   int64(os.Getpid()),
 		OwnerStart: selfProcStart(),
 	}
 	ledger.Leases = append(ledger.Leases, lease)
